% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_text.R
\name{format_text}
\alias{format_text}
\title{Format text into a vector of words}
\usage{
format_text()
}
\value{
A character vector of words extracted from the file
}
\description{
Reads a .txt file from the user's computer using a file selection dialog,
then cleans and converts the text into a vector of individual words.
}
\examples{
\dontrun{
string_vector <- format_text()
}

}
