% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_counts.R
\name{sentiment_counts}
\alias{sentiment_counts}
\title{Count positive and negative words in text}
\usage{
sentiment_counts(words)
}
\arguments{
\item{words}{A character vector of words created using format_text()}
}
\value{
A list containing the number of positive and negative words
}
\description{
Calculates how many words in a cleaned text vector match
predefined positive and negative sentiment lexicons.
}
\examples{
\dontrun{
pos_neg_count <- sentiment_counts(string_vector)
}
}
